+++ finance/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.0.3] finance
 ### Releases
 - [NPM v0.0.3] https://www.npmjs.com/package/ryan-finance/v/0.0.3 - finance
+
+## 2024-07-15 01:16:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.335.0 (2.370.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.1.0] finance
+### Releases
+- [NPM v0.1.0] https://www.npmjs.com/package/ryan-finance/v/0.1.0 - finance